Continue to fix compilation errors.

+++ cocos/core/geometry/AABB.cpp
@@ -90,7 +90,7 @@ void AABB::merge(const std::vector<cc::Vec3> &points) {
 
 void AABB::merge(const Frustum &frustum) {
     const std::array<Vec3, 8> &vertices = frustum.vertices;
-    for (uint i = 0; i < vertices.max_size(); ++i) {
+    for (size_t i = 0; i < vertices.max_size(); ++i) {
         merge(vertices[i]);
     }
 }
